requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
ollama>=0.4.0
httpx>=0.25.0
python-dotenv>=1.0.0
